py_test/requirements.txt
pytest==8.3.5
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
pycryptodome==3.21.0
